+++ packages/terra-status-view/src/StatusView.jsx
@@ -7,7 +7,7 @@ import styles from './StatusView.module.scss';
 
 const cx = classNames.bind(styles);
 
-const variants = {
+const StatusViewVariants = {
   NODATA: 'no-data',
   NOMATCHINGRESULTS: 'no-matching-results',
   NOTAUTHORIZED: 'not-authorized',
@@ -53,7 +53,7 @@ const propTypes = {
    * `no-matching-results`, `not-authorized`,
    * `error`, or `custom`.
    */
-  variant: PropTypes.string,
+  variant: PropTypes.oneOf(['no-data', 'no-matching-results', 'not-authorized', 'error', 'custom']),
 };
 
 const defaultProps = {
@@ -63,7 +63,7 @@ const defaultProps = {
   isGlyphHidden: false,
   message: undefined,
   title: undefined,
-  variant: variants.ERROR,
+  variant: StatusViewVariants.ERROR,
 };
 
 const contextTypes = {
@@ -117,7 +117,7 @@ class StatusView extends React.Component {
 
     contentWithGlyphHeight = contentHeight + (this.glyphNode ? this.glyphNode.getBoundingClientRect().height : 0);
     // if inner view node height is lesser than the components along with glyph inside the inner view.
-    // and the height difference is very less (i.e 0.0001) then add that comparision value to the
+    // and the height difference is very less (i.e 0.0001) then add that comparison value to the
     // the inner view height.
     if (contentWithGlyphHeight - viewHeight <= 0.0001) {
       newViewHeight += 0.0001;
@@ -164,7 +164,7 @@ class StatusView extends React.Component {
 
     let glyphSection;
     if (!isGlyphHidden && this.state.showGlyph) {
-      if (variant === variants.CUSTOM) {
+      if (variant === StatusViewVariants.CUSTOM) {
         glyphSection = (
           <div className={cx('glyph')} ref={(element) => { this.glyphNode = element; }}>
             {customGlyph}
@@ -199,7 +199,7 @@ class StatusView extends React.Component {
 
     let defaultTitle = title;
     if (!defaultTitle) {
-      defaultTitle = (variant === StatusView.Opts.variants.CUSTOM) ? '' : intl.formatMessage({ id: `Terra.status-view.${variant}` });
+      defaultTitle = (variant === StatusViewVariants.CUSTOM) ? '' : intl.formatMessage({ id: `Terra.status-view.${variant}` });
     }
 
     let dividerSection;
@@ -253,6 +253,5 @@ class StatusView extends React.Component {
 StatusView.propTypes = propTypes;
 StatusView.contextTypes = contextTypes;
 StatusView.defaultProps = defaultProps;
-StatusView.Opts = {};
-StatusView.Opts.variants = variants;
 export default StatusView;
+export { StatusViewVariants };